restructures components
