+++ blocks/three-video-block/Edit.js
@@ -21,7 +21,6 @@ import { more } from "@wordpress/icons";
 
 export default function Edit({ attributes, setAttributes, isSelected }) {
 	const onImageSelect = (imageObject) => {
-		console.log(imageObject);
 		setAttributes({ videoUrl: null });
 		setAttributes({
 			videoUrl: imageObject.url,
@@ -89,7 +88,6 @@ export default function Edit({ attributes, setAttributes, isSelected }) {
 
 	function handleClick(objectURL) {
 		if (objectURL) {
-			console.log("success good job", objectURL);
 			onImageSelect(objectURL);
 		}
 		console.log("fail", objectURL);
